pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["UIika <1lys.rynd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.2"}
pymongo = {extras = ["srv"], version = "^4.6.3"}
pydantic = "^2.7.1"
python-jose = "^3.3.0"
jinja2 = "^3.1.3"
python-multipart = "^0.0.9"
passlib = "^1.7.4"
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
